Introduce a base class for Oneoffixx tests
